code refactor and fix issues
