+++ resources/views/admin/createOrEdit.blade.php
@@ -2,6 +2,12 @@
 
 @section('title', 'Coupons Management')
 @section('page-title', isset($coupon) ? 'Edit Coupon' : 'Create Coupon')
+
+@push('styles')
+<link rel="stylesheet" href="https://code.jquery.com/ui/1.13.2/themes/base/jquery-ui.css">
+<link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
+@endpush
+
 @section('breadcrumb')
 <li class="breadcrumb-item"><a href="{{ route('admin.coupons.index') }}">Coupon Manager</a></li>
 <li class="breadcrumb-item active" aria-current="page">{{ isset($coupon) ? 'Edit Coupon' : 'Create Coupon' }}</li>
@@ -13,10 +19,9 @@
     <form action="{{ isset($coupon) ? route('admin.coupons.update', $coupon->id) : route('admin.coupons.store') }}"
         method="POST" enctype="multipart/form-data" id="couponForm">
         @csrf
-        @if(isset($coupon))
+        @isset($coupon)
         @method('PUT')
-        @endif
-
+        @endisset
         <div class="row">
             <div class="col-8">
                 <!-- card section -->
@@ -55,7 +60,6 @@
                                 </div>
                             </div>
                         </div>
-
                         <div class="row">
                             <div class="col-md-6">
                                 <div class="form-group">
@@ -80,7 +84,6 @@
                                 </div>
                             </div>
                         </div>
-
                         <div class="row">
                             <div class="col-md-6">
                                 <div class="form-group">
@@ -106,29 +109,7 @@
                                 </div>
                             </div>
                         </div>
-
                         <div class="row">
-                            <div class="col-md-6">
-                                <div class="form-group">
-                                    <label>Start Date</label>
-                                    <input type="text" name="start_date" id="start_date" class="form-control"
-                                        value="{{ old('start_date', $coupon->start_date ?? '') }}" placeholder="Coupon start date">
-                                    @error('start_date')
-                                    <div class="text-danger validation-error">{{ $message }}</div>
-                                    @enderror
-                                </div>
-                            </div>
-                            <div class="col-md-6">
-                                <div class="form-group">
-                                    <label>End Date</label>
-                                    <input type="text" name="end_date" id="end_date" class="form-control"
-                                        value="{{ old('end_date', $coupon->end_date ?? '') }}" placeholder="Coupon end date">
-                                    @error('end_date')
-                                    <div class="text-danger validation-error">{{ $message }}</div>
-                                    @enderror
-                                </div>
-                            </div>
-
                             <div class="col-md-12">
                                 <div class="form-group">
                                     <label>Notes</label>
@@ -158,7 +139,7 @@
                         <h4 class="card-title">Coupon Applicability</h4>
                     </div>
                     <div class="card-body">
-                        @if(class_exists('admin\\products\\Models\\Product') && class_exists('admin\\categories\\Models\\Category'))
+                        @if(class_exists('\admin\products\Models\Product::class') && class_exists('\admin\categories\Models\Category::class'))
                         <div class="form-group">
                             <label for="categories">Categories</label>
                             <select name="categories[]" id="categories" multiple class="form-control select2">
@@ -176,7 +157,7 @@
                             </select>
                         </div>
                         @endif
-                        @if(class_exists('admin\\courses\\Models\\Course'))
+                        @if(class_exists('\admin\courses\Models\Course'))
                         <div class="form-group">
                             <label for="courses">Courses</label>
                             <select name="courses[]" id="courses" multiple class="form-control select2">
@@ -186,6 +167,23 @@
                             </select>
                         </div>
                         @endif
+
+                        <div class="form-group">
+                            <label>Start Date</label>
+                            <input type="text" name="start_date" id="start_date" class="form-control"
+                                value="{{ old('start_date', $coupon->start_date ?? '') }}" placeholder="Coupon start date">
+                            @error('start_date')
+                            <div class="text-danger validation-error">{{ $message }}</div>
+                            @enderror
+                        </div>
+                        <div class="form-group">
+                            <label>End Date</label>
+                            <input type="text" name="end_date" id="end_date" class="form-control"
+                                value="{{ old('end_date', $coupon->end_date ?? '') }}" placeholder="Coupon end date">
+                            @error('end_date')
+                            <div class="text-danger validation-error">{{ $message }}</div>
+                            @enderror
+                        </div>
                     </div>
                 </div>
             </div>
@@ -197,193 +195,6 @@
 </div>
 @endsection
 
-@push('styles')
-<link rel="stylesheet" href="https://code.jquery.com/ui/1.13.2/themes/base/jquery-ui.css">
-<link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
-@endpush
-
 @push('scripts')
-<script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
-<script src="https://code.jquery.com/ui/1.13.2/jquery-ui.js"></script>
-<script>
-    $(document).ready(function() {
-        ['#courses', '#products', '#categories'].forEach(function(selector) {
-            $(selector).select2({
-                placeholder: "Select " + selector.replace('#', ''),
-                allowClear: true,
-                width: '100%'
-            });
-        });
-
-        try {
-            if (typeof $.fn.select2 !== 'undefined') {
-                $('.select2').select2();
-            }
-        } catch (error) {
-            console.log('Select2 error:', error);
-        }
-
-        // Datepicker for start_date and end_date
-        function customizeDatepickerButtons(input, inst) {
-            setTimeout(function() {
-                var buttonPane = $(inst.dpDiv).find('.ui-datepicker-buttonpane');
-                buttonPane.find('.ui-datepicker-close').hide();
-                buttonPane.find('.ui-datepicker-current').hide();
-                buttonPane.find('.ui-datepicker-clear').remove();
-                $('<button type="button" class="ui-datepicker-clear ui-state-default ui-priority-primary ui-corner-all" style="margin-left:8px;">Clear</button>')
-                    .appendTo(buttonPane)
-                    .on('click', function() {
-                        $(input).val('');
-                        $(input).datepicker('hide');
-                    });
-            }, 1);
-        }
-        var today = new Date();
-        var yyyy = today.getFullYear();
-        var mm = String(today.getMonth() + 1).padStart(2, '0');
-        var dd = String(today.getDate()).padStart(2, '0');
-        var minDate = yyyy + '-' + mm + '-' + dd;
-        $('#start_date, #end_date').datepicker({
-            dateFormat: 'yy-mm-dd',
-            changeMonth: true,
-            changeYear: true,
-            showButtonPanel: true,
-            autoclose: true,
-            minDate: minDate,
-            beforeShow: customizeDatepickerButtons,
-            onChangeMonthYear: function(year, month, inst) {
-                var input = inst.input ? inst.input[0] : null;
-                if (input) customizeDatepickerButtons(input, inst);
-            },
-            onSelect: function(dateText, inst) {
-                var input = inst.input ? inst.input[0] : null;
-                if (input) customizeDatepickerButtons(input, inst);
-            }
-        });
-        $('#start_date, #end_date').on('focus click', function() {
-            $(this).datepicker('show');
-        });
-
-        // Helper functions for error display
-        function showError(inputId, message) {
-            const input = $('#' + inputId);
-            const errorDiv = input.next('.client-validation-error');
-            if (errorDiv.length) {
-                errorDiv.text(message);
-            } else {
-                input.after('<div class="text-danger client-validation-error mt-1" style="font-size: 12px;">' + message + '</div>');
-            }
-        }
-
-        function clearError(inputId) {
-            const input = $('#' + inputId);
-            input.next('.client-validation-error').remove();
-        }
-
-        // Real-time validation for Coupon Code
-        $('#code').on('blur keyup', function() {
-            const value = $(this).val().trim();
-            if (value === '') {
-                showError('code', 'Coupon code is required.');
-            } else if (value.length < 2) {
-                showError('code', 'Coupon code must be at least 2 characters.');
-            } else if (value.length > 50) {
-                showError('code', 'Coupon code must not exceed 50 characters.');
-            } else {
-                clearError('code');
-            }
-        });
-
-        // Real-time validation for Discount Type
-        $('#type').on('blur change', function() {
-            const value = $(this).val();
-            if (value === '') {
-                showError('type', 'Discount type is required.');
-            } else {
-                clearError('type');
-            }
-        });
-
-        // Real-time validation for Discount Value
-        $('#amount').on('blur keyup', function() {
-            const value = $(this).val();
-            if (value === '') {
-                showError('amount', 'Discount value is required.');
-            } else if (parseFloat(value) < 0) {
-                showError('amount', 'Discount value must be at least 0.');
-            } else {
-                clearError('amount');
-            }
-        });
-
-        // Real-time validation for Maximum Uses
-        $('#max_uses').on('blur keyup', function() {
-            const value = $(this).val();
-            if (value !== '' && parseInt(value) < 1) {
-                showError('max_uses', 'Maximum uses must be at least 1.');
-            } else {
-                clearError('max_uses');
-            }
-        });
-
-        // Real-time validation for Status
-        $('#status').on('blur change', function() {
-            const value = $(this).val();
-            if (value === '') {
-                showError('status', 'Status is required.');
-            } else {
-                clearError('status');
-            }
-        });
-
-        // Real-time validation for Notes
-        $('#notes').on('blur keyup', function() {
-            const value = $(this).val().trim();
-            if (value.length > 500) {
-                showError('notes', 'Notes must not exceed 500 characters.');
-            } else {
-                clearError('notes');
-            }
-        });
-
-        // Form submission validation
-        $('#couponForm').on('submit', function(e) {
-            $('.client-validation-error').remove();
-            var hasErrors = false;
-            if ($('#code').val().trim() === '') {
-                showError('code', 'Coupon code is required.');
-                hasErrors = true;
-            }
-            if ($('#type').val() === '') {
-                showError('type', 'Discount type is required.');
-                hasErrors = true;
-            }
-            if ($('#amount').val() === '') {
-                showError('amount', 'Discount value is required.');
-                hasErrors = true;
-            }
-            if ($('#status').val() === '') {
-                showError('status', 'Status is required.');
-                hasErrors = true;
-            }
-            if ($('#start_date').val() !== '' && !/^\d{4}-\d{2}-\d{2}$/.test($('#start_date').val())) {
-                showError('start_date', 'Start date format must be YYYY-MM-DD.');
-                hasErrors = true;
-            }
-            if ($('#end_date').val() !== '' && !/^\d{4}-\d{2}-\d{2}$/.test($('#end_date').val())) {
-                showError('end_date', 'End date format must be YYYY-MM-DD.');
-                hasErrors = true;
-            }
-            if (hasErrors) {
-                e.preventDefault();
-                $('html, body').animate({
-                    scrollTop: $('.client-validation-error').first().offset().top - 100
-                }, 500);
-                return false;
-            }
-            $('#saveBtn').prop('disabled', true)
-                .html('<i class="mdi mdi-loading mdi-spin"></i> Saving...');
-        });
-    });
-</script>
+@include('coupons::admin.partials.script')
 @endpush
